fix: BIGINT-Typ für alle FK-Spalten auf fire_dept.id (errno 150)
MariaDB errno 150 tritt auf, wenn INT-Spalten als FK auf BIGINT-PKs
verweisen. fire_dept.id ist seit Migration 0001 BIGINT; alle
Multi-Tenancy-Migrationen (0044–0054) sowie ORM-Modelle nutzten
fälschlicherweise Integer/INT.

- Migrationen 0044–0054: org_id / *_org_id INT → BIGINT
- Migration 0044 zusätzlich idempotent gemacht (partieller Zustand
  nach fehlgeschlagenem Lauf wird automatisch bereinigt)
- ORM-Modelle: alle fire_dept.id-FKs auf BigInteger korrigiert
  (master, user, incident, invitation, lagekarte, major_incident)
- MIGRATION_RUNBOOK.md: BIGINT-Regel + Idempotenz-Muster dokumentiert
- Wiki + README: abgeschlossene Doku-Updates aus vorheriger Session

+++ alembic/versions/0045_multitenancy_pr2_expand.py
@@ -48,12 +48,13 @@ def upgrade():
         pass
 
     # 3. alarm_type: PK von code auf id (BigInt AUTO_INCREMENT) umstellen + org_id hinzufügen
+    #    org_id BIGINT – muss zu fire_dept.id BIGINT passen (errno 150 sonst)
     conn.execute(text("""
         ALTER TABLE `alarm_type`
           DROP PRIMARY KEY,
           ADD COLUMN `id` BIGINT NOT NULL AUTO_INCREMENT FIRST,
           ADD PRIMARY KEY (`id`),
-          ADD COLUMN `org_id` INT NULL DEFAULT NULL,
+          ADD COLUMN `org_id` BIGINT NULL DEFAULT NULL,
           ADD INDEX `ix_alarm_type_org_id` (`org_id`)
     """))
